Ex1-linear regression/requirements.txt
Packages:
typing
linear_regression
numpy
matplotlib.pyplot
seaborn
pandas
sklearn.model_selection
polynomial_fitting